Process updated and new CRAN packages

Packages which got removed from CRAN can be deleted by setting prune = TRUE. Argument interval allows to specify a range which should be processed.

Usage

process_cran_updates(
  package_name,
  tag,
  platform,
  local_clone_dir,
  interval = lubridate::today(),
  codename = NULL,
  local_build_root = ".",
  endpoint = "https://s3.eu-central-003.backblazeb2.com",
  region = "eu-central-003",
  bucket = "devxy-arm64-r-binaries",
  process_updated = TRUE,
  process_new = TRUE,
  process_removed = TRUE
)

Arguments

package_name

(character)
Package name

tag

(character)
Tag/version

platform

(character)
Platform identifier

local_clone_dir

(character)
Path to clone git repos into

interval

(lubridate::Interval)
Date interval

codename

(character)
Linux distribution identifier

local_build_root

(character)
Path to local build root

endpoint

(character)
S3 endpoint

region

(character)
S3 region

bucket

(character)
S3 bucket name

process_updated

(logical)
Whether to process updated packages on CRAN

process_new

(logical)
Whether to process newly added packages on CRAN

process_removed

(logical)
Whether to process packages that have been removed from CRAN

Examples

if (FALSE) { # \dontrun{
process_cran_updates(
  interval = lubridate::interval(lubridate::today() - 2, lubridate::today() - 20),
  process_updated = FALSE, process_new = FALSE
)
} # }